Split client/lib/analytics into multiple modules.
Instead of one monolithic file that can't easily be tree-shaken, this
change splits `client/lib/analytics` into individual modules.
