+++ client/lib/analytics/utils/is-ad-tracking-allowed.js
@@ -1,0 +1,32 @@
+/**
+ * Internal dependencies
+ */
+import config from 'config';
+import debug from './debug';
+import doNotTrack from './do-not-track';
+import isUrlBlacklistedForPerformance from './is-url-blacklisted-for-performance';
+import isPiiUrl from './is-pii-url';
+import mayWeTrackCurrentUserGdpr from './may-we-track-current-user-gdpr';
+
+/**
+ * Returns whether ad tracking is allowed.
+ *
+ * This function returns false if:
+ *
+ * 1. 'ad-tracking' is disabled
+ * 2. `Do Not Track` is enabled
+ * 3. the current user could be in the GDPR zone and hasn't consented to tracking
+ * 4. `document.location.href` may contain personally identifiable information
+ *
+ * @returns {Boolean} Is ad tracking is allowed?
+ */
+export default function isAdTrackingAllowed() {
+	const result =
+		config.isEnabled( 'ad-tracking' ) &&
+		! doNotTrack() &&
+		! isUrlBlacklistedForPerformance() &&
+		! isPiiUrl() &&
+		mayWeTrackCurrentUserGdpr();
+	debug( `isAdTrackingAllowed: ${ result }` );
+	return result;
+}
