+++ client/lib/analytics/utils/marketing-coupons-key.js
@@ -1,0 +1,7 @@
+/**
+ * `localStorage` key for marketing coupons.
+ *
+ * @type {String} `localStorage` key.
+ */
+const MARKETING_COUPONS_KEY = 'marketing-coupons';
+export default MARKETING_COUPONS_KEY;
