+++ client/lib/analytics/utils/should-report-omit-blog-id.js
@@ -1,0 +1,17 @@
+const SITE_FRAGMENT_REGEX = /\/(:site|:site_id|:siteid|:blogid|:blog_id|:siteslug)(\/|$|\?)/i;
+
+/**
+ * Check if a path should report the currently selected site ID.
+ *
+ * Some paths should never report it because it's used
+ * to tell general admin and site-specific activities apart.
+ *
+ * @param {String} path The tracked path.
+ * @returns {Boolean} If the report should null `blog_id`.
+ */
+export default path => {
+	if ( ! path ) {
+		return true;
+	}
+	return ! SITE_FRAGMENT_REGEX.test( path );
+};
